requirements.txt
g2p-en >= 2.1.0
inflect >= 5.0.0
librosa >= 0.9.2
parallel-wavegan >= 0.5.0
matplotlib >= 3.5.0
numba >= 0.56.0
numpy >= 1.24.0
pycwt >= 0.3.0
pypinyin >= 0.48.0
pyworld >= 0.3.0
PyYAML >= 6.0
scikit-learn >= 1.0.0
scipy >= 1.9.0
setuptools >= 59.0
soundfile >= 0.12.0
tensorboard >= 2.11.0
tgt >= 1.4.4
torch >= 2.0.0
tqdm >= 4.65.0
unidecode >= 1.3.0